+++ doc/source/changelog.rst
@@ -5,7 +5,7 @@ Change Log
 v1.1.0
 ======
 
-* `Support function pre and post hooks to be run prior to and after given commands. <https://github.com/bckohan/django-routines/issues/9>`_
+* `Invalidate importlib caches if command is makemigrations. <https://github.com/bckohan/django-routines/issues/13>`_
 * `Command type for system commands (i.e. non-management commands) to be run as subprocesses <https://github.com/bckohan/django-routines/issues/7>`_
 * `Option to run management commands as subprocesses instead of in the same process space. <https://github.com/bckohan/django-routines/issues/6>`_